requirements.txt
backtrader==1.9.76.123
stable-baselines3==2.3.2
gymnasium==0.29.1
pandas-ta==0.4.71b0
vaderSentiment==3.3.2
ccxt==4.3.47
numpy>=2.2.6
pandas>=2.3.2
python-dotenv==1.0.0
newsapi-python==0.2.7
redis==5.0.1
pytest==8.2.0
matplotlib>=3.7.0
seaborn>=0.13.0
torch>=2.0.0
# Убраны проблемные зависимости:
# TA-Lib==0.4.28  # удалить
# gym==0.21.0     # удалить
